Active state for selected row & for plus
