+++ package.json
@@ -25,6 +25,7 @@
     "body-parser": "1.7.0",
     "browserify": "^10.2.4",
     "browserify-notify": "^1.1.2",
+    "classnames": "^2.1.2",
     "cookie-parser": "1.3.2",
     "domready": "^1.0.8",
     "ejs": "1.0.0",
